requirements.txt
appnope==0.1.3
asgiref==3.7.2
asttokens==2.2.1
attrs==23.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
bs4==0.0.2
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
comm==0.1.3
contourpy==1.1.0
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
Django==5.0.3
et-xmlfile==1.1.0
executing==1.2.0
Flask==3.0.3
Flask-Cors==5.0.0
fonttools==4.41.0
google-api-core==2.18.0
google-api-python-client==2.123.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.63.0
h11==0.14.0
httplib2==0.22.0
idna==3.6
ipykernel==6.24.0
ipython==8.14.0
itsdangerous==2.2.0
jedi==0.18.2
Jinja2==3.1.4
joblib==1.3.1
jupyter_client==8.3.0
jupyter_core==5.3.1
kagglehub==0.3.3
kiwisolver==1.4.4
MarkupSafe==3.0.2
matplotlib==3.7.2
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.25.1
oauthlib==3.2.2
openpyxl==3.1.2
outcome==1.3.0.post0
packaging==23.1
pandas==2.0.3
parso==0.8.3
patsy==0.5.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.9.1
prompt-toolkit==3.0.39
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
Pygments==2.15.1
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3
pyzmq==25.1.0
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
scikit-learn==1.3.0
scipy==1.11.1
selenium==4.17.2
six==1.16.0
sklearn==0.0.post7
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
stack-data==0.6.2
statsmodels==0.14.0
threadpoolctl==3.2.0
tornado==6.3.2
tqdm==4.66.6
traitlets==5.9.0
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.9.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.2.0
wcwidth==0.2.6
Werkzeug==3.1.1
wsproto==1.2.0
